Start with WorkflowOrchestrator class by moving select_template
