+++ test/unit/util/workflows/workflow_test.py
@@ -17,9 +17,7 @@ from exasol.toolbox.util.workflows.templates import (
     WORKFLOW_TEMPLATE_OPTIONS,
 )
 from exasol.toolbox.util.workflows.workflow import (
-    ALL,
     Workflow,
-    _select_workflow_template,
     update_workflow,
 )
 
@@ -231,19 +229,6 @@ class TestWorkflow:
                     output_directory=tmp_path,
                     github_template_dict=project_config.github_template_dict,
                 )
-
-
-class TestSelectWorkflowTemplate:
-    @staticmethod
-    def test_for_all_works_as_expected():
-        result = _select_workflow_template(ALL)
-        assert result == WORKFLOW_TEMPLATE_OPTIONS
-
-    @staticmethod
-    @pytest.mark.parametrize("workflow_name", WORKFLOW_TEMPLATE_OPTIONS)
-    def test_for_individual_workflows_works_as_expected(workflow_name):
-        result = _select_workflow_template(workflow_name)
-        assert result == {workflow_name: WORKFLOW_TEMPLATE_OPTIONS[workflow_name]}
 
 
 class TestUpdateWorkflow:
